dedup build.rs
